Add import statement (#18)
As we discussed on Discord, I am now submitting an initial version of
the `import` statement documentation. All suggestions are welcome.

Fixes: https://github.com/WerWolv/Documentation/issues/15.
Fixes: https://github.com/WerWolv/Documentation/issues/11.

I am not sure if this is considered controversial, but adding a folder
with multiple of your pattern files to the search paths using `Extras >
Settings > Folders` will allow the user to correctly `import` their
pattern files to other pattern files. If I understand the request/issue
correctly, that's also the point of the issue:
https://github.com/WerWolv/ImHex-Patterns/issues/269

**Edit**: I think it also fixes:
https://github.com/WerWolv/ImHex/issues/1641 since ImHex won't crash
anymore in that issue, and this PR explains how to properly use standard
library headers.

---------

Co-authored-by: paxcut <53811119+paxcut@users.noreply.github.com>