More noob-friendly for-loop documentation (#22)
I rewrote and added to the documentation on for loops to make it more
friendly to people who are new to programming, reminiscent of the
resources I was given in the programming classes I've taken.

I found that a lot of the documentation is not very easy to understand
for people who aren't familiar with the languages that Pattern Language
is based on, especially with the library functions. I figured that
helping with what I do understand so far would be a good pursuit.